version: calculate NMA_API_VERSION based on the version number
The concept of NMA_VERSION_CUR_STABLE and NMA_VERSION_NEXT_STABLE is
flawed. Only the current API version of the source matters.

If the current checkout is still in development (from master branch,
odd numbers, e.g. 1.8.29+), then this is already aspiring the upcoming
API (e.g. 1.8.30). Hence, also for such development version, it should
pretend to be already the to-be-released version.

For releases (tagged, with even numbers, like 1.8.28), obviously the
version is exactly the API version.

Since we have a versioning scheme based on odd/even numbers, we can
calculate the current/upcoming API version. Do that.

This also allows us to skip bumping the NMA_VERSION_CUR_STABLE
and NMA_VERSION_NEXT_STABLE macros. That was easy for forget, resulting
in a broken release. While the release process was documented to bump
these two defines, it's easy to forget.

The same is done by libnm, which also follows the odd/even versioning
scheme.