vpn-password-dialog: don't respond to activation with activate_default()
In absence of an widget that receives the default activations this
results in activation of the currently focused widget. This results in
infinite recursion if we're responding to its activation in the first
place.

Even though this would be a programming error, it's difficult to spot
and can be caused just with an ommission of a single property in the
GtkBuilder side.

Let's be on the safe side and just invoke the response directly.