eap-tls: do not set empty 802.1x domain_suffix_match
If the user doesn't fill this in (hopefully knowing what are they
doing), we're setting the property to a blank string. That is no good,
because that results in a connection that can not be activated.

Current version of NetworkManager just passes it as it is to the
supplicant and supplicant rejects it outright (with a fairly unhelpful
"invalid network" message).

While this might warrant a fix in NM, let's make sure we don't produce
such connections with any NM version.

PEAP and TTLS do the right thing already.