cert-chooser-button: fix detection of selected item
Currently, 'changed' returns with no action if the first item
is selected. This commit changes to return with no action only
if there is no action selected in the menu.

In the end of the call, it assigns the selected item of the menu
for the first item of the list. This commit changes for, after
executing the actions upon the selected item, on the cleanup
phase setting the selected item to none again.

https://gitlab.gnome.org/GNOME/libnma/-/merge_requests/39