build: unbreak build with gcr-3.92
Newer gcr pre-release removed the widgets and therefore the library is
not reliant on Gtk anymore, no matter which version. That's good -- we
were only using the GcrCertificate, not the widgets.

Details here:
https://gitlab.gnome.org/GNOME/gcr/-/issues/100

In fact, we weren't even needing the UI part of gcr-3, just the
gcr-base-3 part. That would've not conflicted with Gtk4.

This allows use to simplify things considerably: both gcr-base-3 and
gcr-4 are fine for use for either Gtk3 or Gtk4.

https://gitlab.gnome.org/GNOME/libnma/-/merge_requests/43