mobile-providers: fix path for iso_3166 translations
We were looking for them in /usr/locale instead of /usr/share/locale.

That said, /usr/share/locale still might not be an entirely correct path
for the iso-codes installation, because the iso-codes files are looked
for in other directories than the global /usr.

Let's look for the translations in the same prefix as the xml files
themselves.