mobile-providers: hardwire Kosovo
We use the iso-codes database to get a nice country name for a level 1
(country) alpha-2 (two-letter) code used in the mobile-broadband-database.
This generally works well, as the areas where mobile providers operate
generally have ISO 3166-1 codes assigned. That is, all but Kosovo.

The mobile-broadband-database uses the "user-assigned" "xk" code for
Kosovo, like many countries and international organizations alike use.
This, unfortunatelly, means that we can't use iso-codes to get the full
name.

We could, in theory, use the level 2 code "rs-km", which places the area
inside the Republic of Serbia. That would not affect how we structure
things, but would require us to load the level 2 (districts) database that
is considerably larger than level 1 (countries). Also, the naming in the
iso-codes database is a little confusing at the moment (see
https://salsa.debian.org/iso-codes-team/iso-codes/-/merge_requests/29).

Let's just make sure Kosovo is always there until ISO makes its mind up.