ui-utils: avoid unnecessary filling password
Consumers of libnma (gnome-control-center and nm-connection-editor)
re-fill a connections settings back into the UI, whenever a "changed"
signal gets emitted and the change results in a valid configuration.
Thus we need to carefully avoid triggering more "changed" signals
while (automatically) filling data into the password field.

In gtk3, gtk_editable_set_text took care of it automatically. Here, it
doesn't trigger a signal if the new text is equal to the old text.

In gtk4 however, the "changed" signal is triggered even when the old and
the new text match. Therefore we add explicit checks before setting the
text in the password field to account for this updated behaviour.

https://gitlab.gnome.org/GNOME/gnome-control-center/-/issues/2109
https://gitlab.gnome.org/GNOME/libnma/-/merge_requests/52