wifi-dialog: Checks and logging cleanup
This drops internal usage of g_return_if_fail() as it is meant to check
invariants at API bounds only and not for the regular code flow. This
also removes checks for impossible errors and use more appropriate log
levels, e.g. we shouldn't use warnings to report empty states.

https://gitlab.gnome.org/GNOME/libnma/-/merge_requests/64