wifi-dialog: Only set device in the combo ::changed callback
Register the callback first, so that it will fire up when
gtk_combo_box_set_active() is called. Alter the callback to not expect
the priv->device to be set already.

This avoids a duplicated gtk_tree_model_get() call, but most importantly
prepares things for when signals are hooked earlier, in a template (next
commit).